np-47035: continuing refactoring NvaProjectBuilder
