+++ cristin-project/src/test/java/no/unit/nva/cristin/projects/RandomProjectDataGenerator.java
@@ -34,8 +34,8 @@ import static no.unit.nva.cristin.model.Constants.ORGANIZATION_PATH;
 import static no.unit.nva.cristin.model.Constants.PERSON_PATH_NVA;
 import static no.unit.nva.cristin.model.Constants.PROJECT_PATH_NVA;
 import static no.unit.nva.cristin.model.Constants.CRISTIN_IDENTIFIER_TYPE;
+import static no.unit.nva.cristin.projects.model.cristin.adapter.CristinFundingSourceToFunding.FUNDING_SOURCES;
 import static no.unit.nva.cristin.projects.model.nva.Funding.UNCONFIRMED_FUNDING;
-import static no.unit.nva.cristin.projects.model.nva.NvaProjectBuilder.FUNDING_SOURCES;
 import static no.unit.nva.cristin.projects.model.nva.NvaProjectBuilder.PROJECT_TYPE;
 import static no.unit.nva.cristin.model.Constants.TYPE;
 import static no.unit.nva.cristin.model.Constants.VALUE;
